+++ client/src/components/Header/index.tsx
@@ -1,4 +1,5 @@
 import { Link } from 'react-router-dom'
+
 import styles from './style.module.scss'
 
 const pages: any = {
