+++ client/src/components/LoginPage/index.tsx
@@ -1,7 +1,7 @@
 import React, { useContext, useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
+
 import { AuthContext } from '../../contexts/AuthContext'
-import { http } from '../../utils/http'
 
 import styles from './style.module.scss'
