+++ client/src/components/index.ts
@@ -2,5 +2,9 @@ export { LoginPage } from './LoginPage'
 export { RandomUsersPage } from './RandomUsersPage'
 export { HttpCatPage } from './HttpCatPage'
 export { RandomDogPage } from './RandomDogPage'
-export { Header } from './Header'
 export { ClientsRegistrationPage } from './ClientsRegistrationPage'
+
+export { Header } from './Header'
+export { Pagination } from './Pagination'
+export { SearchInput } from './SearchInput'
+export { ReferenceLink } from './ReferenceLink'
